feat: add more test and check tests
