+++ .github/workflows/ci.yml
@@ -15,10 +15,12 @@ jobs:
   tests:
     name: Tests
     runs-on: ubuntu-latest
+    env:
+      JWT_SECRET_KEY: ${{ secrets.JWT_SECRET_KEY }}
     steps:
       - name: Checkout
         uses: actions/checkout@v4
-
+      
       - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@v3
 
@@ -37,15 +39,18 @@ jobs:
 
       - name: Start services
         run: docker compose up --wait --no-build
+      
+      #      - name: Check HTTP reachability
+      #        run: curl -v --fail-with-body http://localhost
+      #
+      #      - name: Check API reachability
+      #        run: curl -vk --fail-with-body https://localhost
+      #
+      #      - name: Check Frontend reachability
+      #        run: "curl -vk --fail-with-body -H 'Accept: text/html' https://localhost"
 
-      - name: Check HTTP reachability
-        run: curl -v --fail-with-body http://localhost
-
-      - name: Check API reachability
-        run: curl -vk --fail-with-body https://localhost
-
-      - name: Check Frontend reachability
-        run: "curl -vk --fail-with-body -H 'Accept: text/html' https://localhost"
+      - name: Create KEY files for JWT
+        run: docker compose exec -T api bin/console lexik:jwt:generate-keypair --skip-if-exists
 
       - name: Create test database
         run: docker compose exec -T api bin/console -e test doctrine:database:create
@@ -58,6 +63,7 @@ jobs:
 
       - name: Doctrine Schema Validator
         run: docker compose exec -T api bin/console -e test doctrine:schema:validate
+
   lint:
     name: Docker Lint
     runs-on: ubuntu-latest
