+++ api/.gitignore
@@ -22,4 +22,5 @@
 
 ###> lexik/jwt-authentication-bundle ###
 /config/jwt/*.pem
+!/config/jwt/*-test.pem
 ###< lexik/jwt-authentication-bundle ###
