+++ api/config/jwt/private-test.pem
@@ -1,0 +1,30 @@
+-----BEGIN ENCRYPTED PRIVATE KEY-----
+MIIFHDBOBgkqhkiG9w0BBQ0wQTApBgkqhkiG9w0BBQwwHAQIHQPsFX8gboECAggA
+MAwGCCqGSIb3DQIJBQAwFAYIKoZIhvcNAwcECC0tgUSG5Gv0BIIEyDCkXN9Ph+1n
+H/iQuseGaYSNBBOktNK2FLQPPss+w/jtE0FEKQhyih/8RzwpwAFg2M8d0xoux+GM
+Hb3FJkSG8vgf9KtYn+7f0CH6bhU/p305k/76JfSL4/KPheClAAsLpfuzCMIwVakI
+svYLvMJm4iNJXHEgnIHlePodM9Q+uduDMBQanJRHtbeFp0uRyflxuAaOEmkh7tKG
+AuRC29+Gd5iauLvItkZk2OFnZ/ERZDvSMxPpBmhXWnpOB8Z4u2AtiuppQHGin2xt
+9FIuEBHfMUN3JEjFH7s6syJF3NmUiUC70lUck9F0kSCuH+UVxUD23l09Xtjyy8BD
+VZT0N7tgS2lAesuS5ol2kwCLHNg3vHOXphOYhW7dutB9qWlDMIL4cdagcvQkNqDe
+PMuoQFumw4LNgF/rATk2kbmd2pN6I4vnFgiKv8QZSuNJb1gRCb73hpZI7pEhourM
+Se0JYWk9Fni+twSwKdPSljlVFBicH0jTt5QWg8D3hpFyCKhFaxaIF0D+QMvtkduF
+cdfDi8b+ET+PNupaHNlIIkdo9XNcW3QkW5hMeK/oKNZdQfSN6FU5rqma7kIpomfO
+HQHbcCiELsrCifDnJoVsc0Z412lX0xMqBI9qC96bExPvjp7y5crlCac7/EwAtxqd
+nZVkSacyCcUXlo9Z5DfomShqIrKxYjQinGlexr7rUjUSn7v0vrMZQNijtKdqNE4C
+AZl8N4c8eupd4938RibbEw3ZULrHGC6cnrbxYgKm45knGDV+mMCYS7zMn0MyltM9
+4+y1VIE+va8+nkf2rCOqZWwHXw1ci+5IbPI4VEngMYvtchgyeHYG+rPOMpDIVrXE
+gjcNj6tUHttKI2O2pIujajaWZtKSFaJbgYy/DucHZ9AX+Ax4h7KXbEocHqKHeiOj
+7o5HgRZg2+KbaVxoT3B2dxKRTpKM3koXOUVQ4/9z0I2T+pkFETJDQ2qyPtZTqR59
+7W+E2rTJlBw8/wt6lFyHeMpkI1ntzM6HO5RMniH3k5olvrnXOWmYmXGGG686opfa
+yUbUw7pFOR4Jlx7PrSe3wfdze7695H5qIJxhtBJNDKYZ+qsQbjM3DCKyZeRUgXFX
+Ukh3vht9ZNt2rt4YXqhswdKuHnGMtk7JjmYoE46725VZ4kR5E/+ykX9wJOq/B40e
+YL0pbrh9S0karBLSo5OUu26kNPZtZ5qBahqse68RX41lz9g1ehD1GNGxDF+wiDyG
+lSbtChnki4BsB+tAMgI8juaugBWj9LV7x+VMmq/sQ1TPKqR6qqp197H5D2G9b5ad
+mZd6xEEOsvxTfYvu4pTUBSIRGWsczQqVDVUZBZe0UGAFWJg49WGiDzRFIkbQMXN5
+niu/4jdWPlpiyUgHVlSMDjhwPvrNd86+yNfnmj1xHRXBkvw4Hd9/AKeY/ne0fhZh
+6fzNL7/jPS5nmOD4ciGrrk44eGqdM/cQQirYiVV3mwNb9l68mZ2X0UZHtytjJ4Rs
+RUdqMS91IA3Vn8AXYDcDD894E7tNDOZIuKgp0QYxmToHXNOG+VqzmPAquu41r95s
+vh7AjpK8nNvuqISvP5Gsi9DonVDpdnYfKwxwDg7ZLiZjm6zPTTcBxu5UoZPQri0a
+fp14lJI1tw5slmQowmJqMQ==
+-----END ENCRYPTED PRIVATE KEY-----
