+++ api/config/jwt/public-test.pem
@@ -1,0 +1,9 @@
+-----BEGIN PUBLIC KEY-----
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAlVr5bf6tFOIB3gQ60lk3
+Oe0/WyKowAS+6nYYszhXms504O/rFhVXlUce47NgwUM8+e2wk2tuIgi9Qn/2NW8w
+zh6aDOsA79HGT/88j7J3w1lmJYX//15vy7JK0zWTyxrosjYh/vEONSeiIxLYKpV8
+/zW781Wkr1YcstRvMoJiwy1hLLJeMi99SAfNmyvCcqoOKmgl2q2R6CcuBYiQx22w
+lb0ccnUalDMIRmfJGK+VDgA8ztxNYTL4a36a5QD6ichjcarMogxeyLCXn0se0ojk
+Hbt9LZEj0ovfMUwkzRZa8whXhUqps7c5+gBkIJyGkVD1wu33k5jmfY6wnAdMV812
+OQIDAQAB
+-----END PUBLIC KEY-----
