+++ api/src/User/Doctrine/Extension/UserIsSecureExtension.php
@@ -5,6 +5,7 @@ namespace App\User\Doctrine\Extension;
 use ApiPlatform\Doctrine\Orm\Extension\QueryCollectionExtensionInterface;
 use ApiPlatform\Doctrine\Orm\Util\QueryNameGeneratorInterface;
 use ApiPlatform\Metadata\Operation;
+use App\User\Entity\User;
 use Doctrine\ORM\QueryBuilder;
 use Symfony\Bundle\SecurityBundle\Security;
 
@@ -18,6 +19,10 @@ class UserIsSecureExtension implements QueryCollectionExtensionInterface
 
     public function applyToCollection(QueryBuilder $queryBuilder, QueryNameGeneratorInterface $queryNameGenerator, string $resourceClass, Operation $operation = null, array $context = []): void
     {
+        if ($resourceClass !== User::class) {
+            return;
+        }
+
         if ($this->security->getUser() === null) {
             $queryBuilder->andWhere('1 = 0');
             return;
